Feature/experience cud

+++ src/main/java/com/pard/record_on_be/projects/repo/ProjectsRepo.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProjectsRepo extends JpaRepository<Projects, Long> {
+public interface ProjectsRepo extends JpaRepository<Projects, Integer> {
 }
